FIX BUG: SuCOMObject::call() was passing the DISPATCH_PROPERTYGET flag to IDispatch::Invoke().
The problem is that at least one implementation of IDispatch (Microsoft's CPicture class in oleaut32.dll, used by the Suneido Image class) basically has a branch that says "if ((DISPATCH_PROPERTYGET is set) and (number of arguments > 0)) return DISP_E_BADPARAMCOUNT". In other words, if you try to pass arguments to a method via SuCOMObject::call(), the IDispatch will think you're trying to get a property and fail on the grounds that properties don't have parameters.

git-svn-id: svn://svn.code.sf.net/p/suneido/svn/trunk@1230 8ef86ec9-3220-0410-98d7-ddd3c8784f88