Modify SuCOMobject to support IUnknown interface as a fallback where IDispatch not supported.
The purpose of this change is to give Suneido code a 'handle' on which to call Release() if it succeeds in getting a COM pointer to an IUnknown but the underlying object doesn't support IDispatch. See, for example, the Suneido port of the SuImage CSuneido built-in (currently called 'Image2' in stdlib).

git-svn-id: svn://svn.code.sf.net/p/suneido/svn/trunk@1237 8ef86ec9-3220-0410-98d7-ddd3c8784f88