Simplify and clean up my recent changes to SuCOMobject and bring it into line with jSuneido/JSDI COM implementation.
The main principle is that a COMobject should only ever "own" one interface pointer, whether that is an IUnknown or IDispatch. When you call comobject.Release(), it releases that one underlying interface pointer by calling interface->Release(), or it throws because you already released it. If you CoCreateInstance(), the owned pointer is the one returned from CoCreateInstance(). If the Suneido programmer wraps an existing pointer, the pointer given by the Suneido programmer is now "owned" by the SuCOMobject (but if the query for IDispatch is successful, then the SuCOMobject actually owns the queried-for pointer and releases the IUnknown given by the Suneido programmer).

git-svn-id: svn://svn.code.sf.net/p/suneido/svn/trunk@1251 8ef86ec9-3220-0410-98d7-ddd3c8784f88